almost done idp in client ui

+++ client-ui/src/components/Client/Dashboard/pages/UsersManagement/Alarms/Alarms.js
@@ -1,14 +1,25 @@
-import React, { useState } from 'react';
+import React, {useEffect, useState} from 'react';
 import { DataGrid } from '@material-ui/data-grid';
 import './Alarms.css';
 import api from '../../ApiConnections/site-management-api';
 import Select from 'react-select';
 import Popup from 'reactjs-popup';
+import { useKeycloak } from "@react-keycloak/web";
 
 function Alarms() {
+  const { keycloak, initialized } = useKeycloak();
   const [data, setData] = React.useState([]);
   var [buildings] = React.useState([]);
   const [property_id, setPropertyId] = React.useState(0);
+
+  useEffect(() => {
+    if (!!keycloak.authenticated) {
+      console.log(keycloak.tokenParsed.sub);
+      console.log(keycloak.token);
+      localStorage.setItem('token_id', keycloak.tokenParsed.sub);
+      localStorage.setItem('token', keycloak.token);
+    }
+  }, [keycloak.authenticated]);
 
   const customStyles = {
     option: (provided, state) => ({
@@ -20,12 +31,12 @@ function Alarms() {
   }
   
   const loadData = () => {
-    api.get('/users/1/alarms').then(res => {
+    api.get('/users/' + localStorage.getItem('token_id') + '/alarms', {headers:{'Authorization': `Bearer ${localStorage.getItem('token')}`}}).then(res => {
       console.log("loadData");
       setData(res.data);
       console.log(res.data);
     });
-    api.get('/users/1/properties').then(res => {
+    api.get('/users/' + localStorage.getItem('token_id') + '/properties', {headers:{'Authorization': `Bearer ${localStorage.getItem('token')}`}}).then(res => {
       console.log(res.data);
       (res.data).forEach(element => {
         if (buildings.find(building => building.value === element.id) === undefined) {
@@ -41,7 +52,7 @@ function Alarms() {
   }, []);
 
   const handleDelete = id => {
-    api.delete(`/alarms/${id}`).then(res => {
+    api.delete(`/alarms/${id}`, {headers:{'Authorization': `Bearer ${localStorage.getItem('token')}`}}).then(res => {
       console.log(res.affectedRows);
       setData(data.filter(item => item.id !== id));
       loadData();
@@ -49,7 +60,7 @@ function Alarms() {
   };
 
   const addAlarm = () => {
-    api.post('alarms/?property_id=' + property_id, {}).then(res => {
+    api.post('alarms/?property_id=' + property_id, {}, {headers:{'Authorization': `Bearer ${localStorage.getItem('token')}`}}).then(res => {
       console.log(res.data);
       loadData();
     });
